Internal #861: Window Value Functions
Move the WindowValueExecutor classes to separate files.
