+++ src/function/window/CMakeLists.txt
@@ -5,7 +5,8 @@ add_library_unity(
   window_collection.cpp
   window_executor.cpp
   window_segment_tree.cpp
-  window_shared_expressions.cpp)
+  window_shared_expressions.cpp
+  window_value_function.cpp)
 
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_func_window>
