+++ src/include/duckdb/function/window/window_value_function.hpp
@@ -1,0 +1,84 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/function/window/window_value_function.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/function/window/window_executor.hpp"
+
+namespace duckdb {
+
+// Base class for non-aggregate functions that have a payload
+class WindowValueExecutor : public WindowExecutor {
+public:
+	WindowValueExecutor(BoundWindowExpression &wexpr, ClientContext &context, WindowSharedExpressions &shared);
+
+	void Finalize(WindowExecutorGlobalState &gstate, WindowExecutorLocalState &lstate,
+	              CollectionPtr collection) const override;
+
+	unique_ptr<WindowExecutorGlobalState> GetGlobalState(const idx_t payload_count, const ValidityMask &partition_mask,
+	                                                     const ValidityMask &order_mask) const override;
+	unique_ptr<WindowExecutorLocalState> GetLocalState(const WindowExecutorGlobalState &gstate) const override;
+
+	//! The column index of the value column
+	column_t child_idx = DConstants::INVALID_INDEX;
+	//! The column index of the Nth column
+	column_t nth_idx = DConstants::INVALID_INDEX;
+	//! The column index of the offset column
+	column_t offset_idx = DConstants::INVALID_INDEX;
+	//! The column index of the default value column
+	column_t default_idx = DConstants::INVALID_INDEX;
+};
+
+//
+class WindowNtileExecutor : public WindowValueExecutor {
+public:
+	WindowNtileExecutor(BoundWindowExpression &wexpr, ClientContext &context, WindowSharedExpressions &shared);
+
+protected:
+	void EvaluateInternal(WindowExecutorGlobalState &gstate, WindowExecutorLocalState &lstate, DataChunk &eval_chunk,
+	                      Vector &result, idx_t count, idx_t row_idx) const override;
+};
+class WindowLeadLagExecutor : public WindowValueExecutor {
+public:
+	WindowLeadLagExecutor(BoundWindowExpression &wexpr, ClientContext &context, WindowSharedExpressions &shared);
+
+	unique_ptr<WindowExecutorLocalState> GetLocalState(const WindowExecutorGlobalState &gstate) const override;
+
+protected:
+	void EvaluateInternal(WindowExecutorGlobalState &gstate, WindowExecutorLocalState &lstate, DataChunk &eval_chunk,
+	                      Vector &result, idx_t count, idx_t row_idx) const override;
+};
+
+class WindowFirstValueExecutor : public WindowValueExecutor {
+public:
+	WindowFirstValueExecutor(BoundWindowExpression &wexpr, ClientContext &context, WindowSharedExpressions &shared);
+
+protected:
+	void EvaluateInternal(WindowExecutorGlobalState &gstate, WindowExecutorLocalState &lstate, DataChunk &eval_chunk,
+	                      Vector &result, idx_t count, idx_t row_idx) const override;
+};
+
+class WindowLastValueExecutor : public WindowValueExecutor {
+public:
+	WindowLastValueExecutor(BoundWindowExpression &wexpr, ClientContext &context, WindowSharedExpressions &shared);
+
+protected:
+	void EvaluateInternal(WindowExecutorGlobalState &gstate, WindowExecutorLocalState &lstate, DataChunk &eval_chunk,
+	                      Vector &result, idx_t count, idx_t row_idx) const override;
+};
+
+class WindowNthValueExecutor : public WindowValueExecutor {
+public:
+	WindowNthValueExecutor(BoundWindowExpression &wexpr, ClientContext &context, WindowSharedExpressions &shared);
+
+protected:
+	void EvaluateInternal(WindowExecutorGlobalState &gstate, WindowExecutorLocalState &lstate, DataChunk &eval_chunk,
+	                      Vector &result, idx_t count, idx_t row_idx) const override;
+};
+
+} // namespace duckdb
